bot/src/requirements.txt
pymongo==3.0.3
geopy==1.11.0
requests==2.3.0
pyowm==2.3.1
python-telegram-bot==4.0rc1
pytz==2015.7
